requirements.txt
# FastAPI for API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain ecosystem
langchain==0.3.7
langchain-core>=0.3.17,<0.4.0
langchain-community==0.3.7
langchain-google-genai>=0.2.0
langchain-huggingface==0.1.0
langchain-chroma==0.1.4

# Google GenAI (new package, replacing deprecated google-generativeai)
google-genai>=0.3.0

# Vector DB
chromadb==0.4.22
faiss-cpu==1.9.0.post1

# ML + Embeddings - UPDATED VERSIONS
sentence-transformers>=2.6.0
transformers>=4.36.0
huggingface-hub>=0.23.0
tokenizers>=0.15.0

# File processing + OCR
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.2.0
pytesseract==0.3.10
python-docx==1.1.0

# MIME magic
python-magic==0.4.27

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
colorama==0.4.6
typing-extensions>=4.8.0
pydantic>=2.7.4,<3.0.0
numpy<2.0.0
